All links and QR feature added, delete feature added

+++ backend/controllers/links_controller.js
@@ -3,49 +3,53 @@ const bcrypt = require('bcrypt');
 const saltRounds = 10;
 const LINK = require('../models/links');
 const USER = require('../models/user');
+const QR = require('../models/qr');
 const insertHit = require('../utils/hitInsert');
 const generateKey = require('../lib/keyGenerater');
 
-function validateOne(payload) {
-    let errors = {};
-    let isFormValid = true;
-
-    if(!payload || typeof payload.title !== 'string' || payload.title.trim().length === 0) {
-        isFormValid = false;
-        errors.title = 'Title Empty or typeof title is not String';
-    }
-
-    if(!payload || typeof payload.url !== 'string' || payload.url.trim().length === 0){
-        isFormValid = false;
-        errors.url = 'URL Empty or typeof URL is not String';
-    }
-
-    if(!payload || typeof payload.key !== 'string') {
-        isFormValid = false;
-        errors.key = 'typeof KEY is not String'
-    }
-
-    return {
-        success : isFormValid,
-        errors
-    }
-}
-
-async function generateRandomString(destinationUrl){
-    let salt = await bcrypt.genSalt(saltRounds);
-    shortUrl = await bcrypt.hash(destinationUrl, salt);
-    return shortUrl;
-}
-
-// ----             TODO
 module.exports.getAllLinksofUser = async function(req, res){
+    let user = await USER.findById(req.user._id).populate({
+        path : 'links',
+        Model : 'Link',
+        select : 'key url title hitCount qrcode',
+        populate : {
+            path : 'qrcode',
+            Model : 'QRCode',
+            select : 'key url title'
+        }
+    })
     return res.status(200).json({
-        message : 'Gotcha'
+        success : true,
+        message : 'Gotcha',
+        data : user.links
     });
+};
+
+module.exports.handleRedirect = async function(req, res){
+    try {
+        const keyFromReq = req.params.id;
+        const link = await LINK.find({
+            key : keyFromReq
+        });
+        if(link.length == 0){
+            return res.redirect('http://localhost:3000');
+        }
+        let ip = req.socket.remoteAddress;
+        let userAgent = req.headers['user-agent'];
+        await insertHit(ip, userAgent, link[0]._id);
+        return res.redirect(link[0].url);
+    } catch(error) {
+        return res.redirect('http://localhost:3000');
+    }
 };
 
 module.exports.createLink = async function(req, res){
     try {
+        let createLinkForm = req.body;
+        let validationResult = validateOne(createLinkForm);
+        if(!(await validationResult).success){
+            throw new Error('Create Link form validation failed');
+        }
         let title = req.body.title;
         let destinationUrl = req.body.url;
         let shortUrl = req.body.key;
@@ -64,6 +68,7 @@ module.exports.createLink = async function(req, res){
         newUser.save();
         
         return res.status(200).json({
+            success : true,
             message : 'link created',
             data : ans
         });
@@ -71,32 +76,10 @@ module.exports.createLink = async function(req, res){
     } catch (error) {
         console.log(error);
         return res.status(400).json({
-            message: 'Something went wrong, please try again.'
+            success : false,
+            message: error.message
         });
     }
-};
-
-/*
-** ---------- To Handle Redirect ----------
-*/
-
-module.exports.handleRedirect = async function(req, res){
-    const keyFromReq = req.params.id;
-    const link = await LINK.find({
-        key : keyFromReq
-    });
-    if(link.length == 0){
-        return res.status(404).json({
-            message : 'Link not found'
-        });
-    }
-    let ip = req.socket.remoteAddress;
-    let userAgent = req.headers['user-agent'];
-    await insertHit(ip, userAgent, link[0]._id);
-    return res.status(200).json({
-        message : 'Link Found',
-        data : link[0].url
-    })
 };
 
 module.exports.edit = async function(req, res){
@@ -119,24 +102,84 @@ module.exports.edit = async function(req, res){
         link.url = req.body.url;
         await link.save();
         return res.status(200).json({
+            success : true,
             message : 'Link updated',
             data : link
         });
     } catch (error) {
         console.log(error);
         return res.status(400).json({
+            success : false,
             message: error.message
         });
     }
 }
-// have to work on this
-// as this link is getting deleted,, we have to go to the user and need to remove this link for users link array
+
 module.exports.delete = async function(req, res){
-    let link = await LINK.deleteOne({
-        _id : req.params.id
-    })
-    return res.status(200).json({
-        message : 'Link Deleted',
-        data : link
-    })
+    try {
+        let userID = req.user._id;
+        let newLink = await LINK.findById(req.params.id);
+        let QRid;
+        if(newLink.qrcode) {
+            QRid = newLink.qrcode.valueOf();
+        }
+        if(userID.valueOf() !== newLink.user._id.valueOf()){
+            throw new Error('Unauthorized Access to delete a Link');
+        }
+        let updatedUser = await USER.findByIdAndUpdate(req.user._id, {
+            $pull : {
+                links : req.params.id,
+                qrcode : QRid
+            }
+        });
+        await QR.findByIdAndDelete(QRid);
+        let link = await LINK.deleteOne({
+            _id : req.params.id
+        });
+        return res.status(200).json({
+            success : true,
+            message : 'Link Deleted',
+            user : updatedUser,
+            link : link
+        })
+    } catch(error){
+        console.log(error);
+        return res.status(400).json({
+            success : false,
+            message: error.message
+        });
+    }
+}
+
+async function generateRandomString(destinationUrl){
+    let salt = await bcrypt.genSalt(saltRounds);
+    shortUrl = await bcrypt.hash(destinationUrl, salt);
+    return shortUrl;
+}
+
+async function validateOne(payload) {
+    let errors = {};
+    let isFormValid = true;
+
+    if(!payload || typeof payload.title !== 'string' || payload.title.trim().length === 0) {
+        console.log(1);
+        isFormValid = false;
+        errors.title = 'Title Empty or typeof title is not String';
+    }
+
+    if(!payload || validator.isURL(payload.url, { require_tld : false }) == false || validator.isEmpty(payload.url)){
+        isFormValid = false;
+        errors.url = 'URL Empty or it is not a url';
+    }
+
+    if(!payload || payload.key && typeof payload.key !== 'string' && payload.key.length() != 10) {
+        console.log(3);
+        isFormValid = false;
+        errors.key = 'Key is not of a valid type'
+    }
+
+    return {
+        success : isFormValid,
+        errors
+    }
 }
